requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
crewai>=0.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
pymongo>=4.0.0
pandas>=2.0.0
jinja2>=3.0.0